+++ bank-app-backend/src/main/kotlin/org/fando/piris/piris/models/RequestContract.kt
@@ -17,5 +17,5 @@ data class RequestContract(
         @JsonProperty("amount")
         val amount: BigDecimal,
         @JsonProperty("percents")
-        val percents: Double
+        val percents: BigDecimal
 )
